Small improvements
Former-commit-id: 3cbf706e961c33bde3b2ac4c29900b58641cff54 [formerly c83c717c075bbd18dd3490e6d078a97e69230fd5]
Former-commit-id: 000365350efd304a2623bdbe542b2abdd492dba2

+++ src/utils.py
@@ -102,8 +102,11 @@ class Rectangle:
         aou = r1.get_area() + r2.get_area() - aoo
         return aoo / aou
 
+
 # Optional type cast helper
 T = TypeVar('T')
+
+
 def assert_type(arg: Optional[T]) -> T:
     assert arg is not None
     return arg
@@ -111,6 +114,7 @@ def assert_type(arg: Optional[T]) -> T:
 
 def get_capture_size(capture: cv2.VideoCapture) -> Tuple[int, int]:
     return (int(capture.get(3)), int(capture.get(4)))
+
 
 def get_output(filename: str, capture: cv2.VideoCapture = None, capture_size: Tuple[int, int] = None, is_grey: bool = False) -> cv2.VideoWriter:
     fourcc = cv2.VideoWriter_fourcc(*'mp4v')
@@ -120,11 +124,14 @@ def get_output(filename: str, capture: cv2.VideoCapture = None, capture_size: Tu
 
     return cv2.VideoWriter(filename, fourcc, 15.0, capture_size, not is_grey)
 
+
 def get_sequence_length(path: str) -> int:
     return count_dir(path)
 
+
 def get_frame_count(cap: cv2.VideoCapture) -> int:
     return int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+
 
 def get_fps(cap: cv2.VideoCapture) -> int:
     return int(cap.get(cv2.CAP_PROP_FPS))
@@ -137,32 +144,41 @@ def get_kitti_path(sequence: str) -> str:
     # pose_path = '{}/data_odometry_poses/dataset/poses/00.txt'.format(kitti_path)
     return img_path
 
+
 def get_kitti_capture(sequence: str) -> cv2.VideoCapture:
     path = get_kitti_path(sequence)
     return cv2.VideoCapture(path + '/%6d.png'), count_dir(path)
 
 # Cenek Albl et al.
+
+
 def get_cenek_path(sequence: str, camera: int) -> Tuple[str, str]:
     cenek_path = os.environ['CENEK_PATH']
     img_path = f'{cenek_path}/{sequence}/{camera}.mp4'
     ann_path = f'{cenek_path}/{sequence}/detections/{camera}.txt'
     return img_path, ann_path
 
+
 def get_cenek_capture(sequence: str, camera: int) -> Tuple[cv2.VideoCapture, int]:
     cap = cv2.VideoCapture(get_cenek_path(sequence, camera)[0])
     return cap, get_frame_count(cap)
 
+
 def get_cenek_annotation(sequence: str, camera: int) -> str:
     return get_cenek_path(sequence, camera)[1]
+
 
 def get_train_capture() -> cv2.VideoCapture:
     path = 'media/train.mp4'
     return cv2.VideoCapture(path), 1e4
 
+
 def count_dir(path: str) -> int:
     return len(os.listdir(path))
 
 # Math utils
+
+
 def line_intersection(line1: Tuple[Tuple[float, float], Tuple[float, float]], line2: Tuple[Tuple[float, float], Tuple[float, float]]) -> Tuple[float, float]:
     xdiff = (line1[0][0] - line1[1][0], line2[0][0] - line2[1][0])
     ydiff = (line1[0][1] - line1[1][1], line2[0][1] - line2[1][1])
@@ -179,8 +195,10 @@ def line_intersection(line1: Tuple[Tuple[float, float], Tuple[float, float]], li
     y = det(d, ydiff) / div
     return x, y
 
+
 def line_angle(diff1: float, diff2: float) -> float:
     return float(np.arccos(np.dot(diff1, diff2) / (np.linalg.norm(diff1) * np.linalg.norm(diff2))))
+
 
 def read_flow(filename: str) -> np.ndarray:
     """Read flow from .flo file
@@ -203,7 +221,8 @@ def read_flow(filename: str) -> np.ndarray:
 
         return np.resize(data, (int(h), int(w), 2))
 
-def write_flow(filename: str, uv: np.ndarray, v: np.ndarray=None) -> None:
+
+def write_flow(filename: str, uv: np.ndarray, v: np.ndarray = None) -> None:
     """ Write optical flow to file.
 
     If v is None, uv is assumed to contain both u and v channels,
@@ -217,13 +236,13 @@ def write_flow(filename: str, uv: np.ndarray, v: np.ndarray=None) -> None:
     if v is None:
         assert(uv.ndim == 3)
         assert(uv.shape[2] == 2)
-        u = uv[:,:,0]
-        v = uv[:,:,1]
+        u = uv[:, :, 0]
+        v = uv[:, :, 1]
     else:
         u = uv
 
     assert(u.shape == v.shape)
-    height,width = u.shape
+    height, width = u.shape
 
     with open(filename, 'wb') as f:
         # write the header
@@ -232,9 +251,10 @@ def write_flow(filename: str, uv: np.ndarray, v: np.ndarray=None) -> None:
         np.array(height).astype(np.int32).tofile(f)
         # arrange into matrix form
         tmp = np.zeros((height, width*nBands))
-        tmp[:,np.arange(width)*2] = u
-        tmp[:,np.arange(width)*2 + 1] = v
+        tmp[:, np.arange(width)*2] = u
+        tmp[:, np.arange(width)*2 + 1] = v
         tmp.astype(np.float32).tofile(f)
+
 
 def blockshaped(arr: np.ndarray, nrows: int, ncols: int) -> np.ndarray:
     """
@@ -250,8 +270,9 @@ def blockshaped(arr: np.ndarray, nrows: int, ncols: int) -> np.ndarray:
     assert h % nrows == 0, "{} rows is not evenly divisble by {}".format(h, nrows)
     assert w % ncols == 0, "{} cols is not evenly divisble by {}".format(w, ncols)
     return (arr.reshape(h//nrows, nrows, -1, ncols)
-               .swapaxes(1,2)
+               .swapaxes(1, 2)
                .reshape(-1, nrows, ncols))
+
 
 def img_to_video(input: str, output: str, framerate: int = 30) -> None:
     """Convert images to a video
@@ -269,6 +290,7 @@ def img_to_video(input: str, output: str, framerate: int = 30) -> None:
         command = f'ffmpeg -start_number {start_number} -r {framerate} -i {input} -c:v libx264 -vf fps={framerate} -pix_fmt yuv420p {output} -y'
         subprocess.call(command.split(' '))
 
+
 def is_rotation_matrix(R: np.ndarray) -> bool:
     """Checks if a matrix is a valid rotation matrix.
 
@@ -280,9 +302,10 @@ def is_rotation_matrix(R: np.ndarray) -> bool:
     """
     Rt = np.transpose(R)
     shouldBeIdentity = np.dot(Rt, R)
-    I = np.identity(3, dtype = R.dtype)
+    I = np.identity(3, dtype=R.dtype)
     n = np.linalg.norm(I - shouldBeIdentity)
     return cast(bool, n < 1e-6)
+
 
 def rotation_matrix_to_euler(R: np.ndarray) -> np.ndarray:
     """
@@ -298,16 +321,16 @@ def rotation_matrix_to_euler(R: np.ndarray) -> np.ndarray:
     """
     assert(is_rotation_matrix(R))
 
-    sy = np.sqrt(R[0,0] * R[0,0] +  R[1,0] * R[1,0])
+    sy = np.sqrt(R[0, 0] * R[0, 0] + R[1, 0] * R[1, 0])
     singular = sy < 1e-6
 
     if not singular:
-        x = np.arctan2(R[2,1] , R[2,2])
-        y = np.arctan2(-R[2,0], sy)
-        z = np.arctan2(R[1,0], R[0,0])
+        x = np.arctan2(R[2, 1], R[2, 2])
+        y = np.arctan2(-R[2, 0], sy)
+        z = np.arctan2(R[1, 0], R[0, 0])
     else:
-        x = np.arctan2(-R[1,2], R[1,1])
-        y = np.arctan2(-R[2,0], sy)
+        x = np.arctan2(-R[1, 2], R[1, 1])
+        y = np.arctan2(-R[2, 0], sy)
         z = 0
 
     return np.rad2deg(np.array([x, y, z]))
